Bubble Sort - 1.0
This is one of the simplest sorting algorithms, I have
made an unoptimized version of the sort which takes
as long as the time it takes even in the case of sorted
array input. I do not intend to further optimize it
because I need to focus on other topics. Fell free to
improve upon the code.